+++ changelog.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Added new alarm code for Nexus Evolution
 - Minor mod to maintenance page update to allow for faster loading
 - Corrected one issue with power log error check
+- added check for python being installed in the install script
 
 ## V1.19.03 -2024-04-19
 - New add on program genmqttin.py. See wiki and add on page for details